+++ .travis.yml
@@ -29,7 +29,6 @@ before_script:
 
 install: 
 - pip install -U tox-travis
-# - pip install -r requirements.txt
 
 script:
 - tox -e $TOX_ENV
