+++ TODO.md
@@ -1,5 +1,5 @@
 
-- [ ] Fix [WinError 10053] ConnectionAbortedError in restricted_document test - Check fixture scope
+- [x] Fix [WinError 10053] ConnectionAbortedError in restricted_document test - Check fixture scope
 - [ ] Change utilities to utils
 - [ ] Add more tests
 - [ ] Mute "DEVTOOLS Listening..."
