+++ tests/test_greater_than_last_page.py
@@ -8,7 +8,7 @@ import sys
 # sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 # from scribd_dl import ScribdDL
 import pytest
-from scribd_dl.utilities import (
+from scribd_dl.utils import (
     valid_url,
     valid_pages,
     GreaterThanLastPageError
